REFACTOR: move code to socket_server
